+++ lib/core/algorithm/random_algorithm.dart
@@ -16,7 +16,7 @@ class RandomAlgorithm extends GraphAlgorithm {
   void compute(Vertex v, Graph graph) {
     super.compute(v, graph);
     if (v.position == Vector2(0, 0)) {
-      var size = v.cpn!.gameRef.camera.visibleWorldRect;
+      var size = v.cpn!.game.camera.visibleWorldRect;
       v.radius = 5;
       v.position = Vector2(math.Random().nextDouble() * (size.width - 50) + 25,
           math.Random().nextDouble() * (size.height - 50) + 25);