+++ lib/widgets/vertex_component.dart
@@ -18,7 +18,7 @@ class VertexComponent extends ShapeComponent
     with
         TapCallbacks,
         HoverCallbacks,
-        HasGameRef<GraphComponent>,
+        HasGameReference<GraphComponent>,
         CollisionCallbacks
     implements SizeProvider {
   late Vertex vertex;
@@ -51,9 +51,9 @@ class VertexComponent extends ShapeComponent
   String get overlayName => 'vertex${vertex.id}';
 
   Duration get panelDelay =>
-      gameRef.options.panelDelay ?? const Duration(milliseconds: 300);
+      game.options.panelDelay ?? const Duration(milliseconds: 300);
 
-  bool get hasPanel => gameRef.options.vertexPanelBuilder != null;
+  bool get hasPanel => game.options.vertexPanelBuilder != null;
 
   @override
   FutureOr<void> onLoad() {
@@ -66,11 +66,11 @@ class VertexComponent extends ShapeComponent
   }
 
   void loadOverlay() {
-    var panelBuilder = gameRef.options.vertexPanelBuilder;
+    var panelBuilder = game.options.vertexPanelBuilder;
     if (!hasPanel) return;
 
-    gameRef.overlays.addEntry(overlayName, (_, game) {
-      return panelBuilder!(vertex, gameRef.camera.viewfinder);
+    game.overlays.addEntry(overlayName, (_, g) {
+      return panelBuilder!(vertex, game.camera.viewfinder);
     });
   }
 
@@ -94,13 +94,17 @@ class VertexComponent extends ShapeComponent
   void render(Canvas canvas) =>
       vertexShape.render(vertex, canvas, paint, paintLayers);
 
-  VertexShape get vertexShape => gameRef.options.vertexShape;
+  VertexShape get vertexShape => game.options.vertexShape;
 
   @override
   void update(double dt) {
     super.update(dt);
+    vertexUpdate();
+  }
+
+  vertexUpdate() {
     size = vertexShape.size(vertex);
-    algorithm.$size.value = Size(gameRef.size.x, gameRef.size.y);
+    algorithm.$size.value = Size(game.size.x, game.size.y);
 
     algorithmCompute(algorithm);
     hitBox?.position = position;
@@ -120,7 +124,7 @@ class VertexComponent extends ShapeComponent
     if (hasPanel) {
       Future.delayed(panelDelay, () {
         if (isHovered) {
-          gameRef.overlays.add(overlayName);
+          game.overlays.add(overlayName);
         }
       });
     }
@@ -131,7 +135,7 @@ class VertexComponent extends ShapeComponent
     graph.hoverVertex = null;
     if (hasPanel) {
       Future.delayed(panelDelay, () {
-        gameRef.overlays.remove(overlayName);
+        game.overlays.remove(overlayName);
       });
     }
   }
@@ -139,8 +143,8 @@ class VertexComponent extends ShapeComponent
   @mustCallSuper
   void onDrag(Vector2 globalDelta) {
     if (hasPanel) {
-      gameRef.overlays.remove(overlayName);
-      gameRef.overlays.add(overlayName);
+      game.overlays.remove(overlayName);
+      game.overlays.add(overlayName);
     }
     algorithm.afterDrag(vertex, globalDelta);
   }
@@ -181,7 +185,7 @@ class VertexComponent extends ShapeComponent
   }
 
   String? displayName() {
-    var txt = gameRef.options.textGetter.call(vertex);
+    var txt = game.options.textGetter.call(vertex);
     return '${vertex.id}${txt != vertex.id ? " ($txt) " : ""}';
   }
 }